[RENAME] postDetail -> postData 로 이름 변경 #NADA-49
