+++ src/Community/components/postDetail/CountComment.js
@@ -1,9 +1,9 @@
 /** CountComment 게시물 답글의 총 갯수를 반환하는 컴포넌트  */
 import { useSelector } from 'react-redux';
-import { PostDetailSelector } from '../../modules/redux/postDetail';
+import { postDataSelector } from '../../modules/redux/postData';
 
 const CountComment = () => {
-  const data = useSelector(PostDetailSelector('PostData', 'data'));
+  const data = useSelector(postDataSelector('data'));
 
   let totalComments = 0;
