+++ src/index.js
@@ -14,7 +14,7 @@ import activitySlice from './Acitivity/modules/activity.js';
 import barSlice from './Bar/modules/redux/bar.js';
 import filterSlice from './modules/redux/filter.js';
 import postWriteSlice from './Community/modules/redux/postWrite.js';
-import PostDetailSlice from './Community/modules/redux/postDetail.js';
+import postDataSlice from './Community/modules/redux/postData.js';
 import CommentSlice from './Community/modules/redux/comment.js';
 import { modalSlice } from './Modal/modules/redux/modal.js';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
@@ -34,7 +34,7 @@ const store = configureStore({
     modal: modalSlice.reducer,
     filter: filterSlice.reducer,
     postwrite: postWriteSlice.reducer,
-    postdetail: PostDetailSlice.reducer,
+    postData: postDataSlice.reducer,
     comment: CommentSlice.reducer,
     search: searchSlice.reducer,
     badge: badgeSlice.reducer,
